Added Hub Level & Compass / Waypoints
Need to add element symbols to doors to make it easy to identify which level is which, but can't find any CC licensed symbols which would be good. D:

Place waypoints where you want to direct the player, make sure you have triggers to activate & deactivate the waypoints.

Need to create & add the compass widget to the viewport through the player script.