Added new widget + Some save file interactions
Changed Confirmation widget to have an extra button to cancel the action. Added a confirmation for loading a save.
Added a widget that will be used when a Save/Load/Delete is completed. If it fails to complete the task it will reflect that.
Can no longer delete a save currently being used. Can no longer load a save that does not exist.